7.6
Функция вернет true если найден подходящая запись в базе и получено подтверждение на доступ(isResource(resource: String, itemResource: String): Boolean)

+++ src/RoleResource.kt
@@ -3,5 +3,5 @@ import models.Roles
 data class RoleResource(
         val role: Roles,
         val resource: String,
-        val idUser: Long,
+        val idUser: Int,
 )
